requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.11
djangorestframework==3.14.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
frozenlist==1.4.1
googlemaps==4.10.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
multidict==6.0.5
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.2.0
promise==2.3
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.6.0
text-unidecode==1.3
twilio==9.0.0
typing-extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
